[Issue #1657] Setup Next-Intl for App Pages (#1851)
## Summary
Fixes #1657 

### Time to review: __30 mins__

## Changes proposed
Enables next-intl to enable translation for app pages. This sets up
next-intl and enables it for the search page and app not found page.
This adds one other app component. There are three which will be removed
for #1361 .

## Context for reviewers
The strings being declared and injected is still ugly and not addressed
with this, but will be fixed with #1361 .

The `[locale]` folder has not been setup yet b/c it took over the full
routing for the site. That folder will be setup in #1361 .