[Issue #1940] Fix factories to initialize sequence based on DB values (#1941)
## Summary
Fixes #1940

### Time to review: __2 mins__

## Changes proposed
Adjusts our factories to handle loading the values from the DB

Removes the now-duplicate code from the db-seed-local

## Context for reviewers
There's currently a bug with the db-seed-local script due to how the
factories setup the primary keys. If you rerun the script, it doesn't
know where to start for the opportunity assistance listing table, and
starts again at `1`. This fixes that by making it start from the correct
DB value.