[Issue #1985] Allow yes/no for transformed booleans (#1986)
## Summary
Fixes #1985

### Time to review: __2 mins__

## Changes proposed
When handling transformations of booleans, allow `Yes` and `No` as valid
values in addition to `Y` and `N`.

## Context for reviewers
These new values are immensely uncommon, but there are a few rare
occurrences of them. Easier to just support them than work around in any
other way.

## Additional information
For example, in `TSYNOPSIS_HIST`, the count of values for the
cost_sharing bool:

```
No	971
N	60820
Y	13723
Yes	581
```
The non-history table does not have this value.