[Issue #2072] Locally, preserve the auth token in the OpenAPI across refreshes (navapbc/simpler-grants-gov#67)
Fixes #2072

Set the `persistAuthorization` OpenAPI config locally to True

For local development, we frequently need to go to
http://localhost:8080/docs - enter the auth token, and then repeat this
process every time we reopen this page or refresh. Having to either copy
paste or retype in the auth token is tedious. This flag makes it so it
gets preserved in your browsers local storage.

We are only enabling this for the local endpoint at the moment as there
are possibly security implications we would need to consider non-locally
(eg. what if someone is using a public computer).