[Issue #2063] Adjust docker commands based on recent updates (navapbc/simpler-grants-gov#89)
Fixes #2063

Remove the version field from the docker-compose files

Adjust the docker commands to use `docker compose` instead of
`docker-compose`

A recent version of docker removed the need for specifying a version in
the compose files - so the field is now obsolete and just gives a
warning whenever you run a command:
https://docs.docker.com/compose/compose-file/04-version-and-name/

The change in the command we run is actually from 2021 and makes sure we
use docker compose v2. Depending on how you've setup your local instance
of docker, `docker-compose` may have been aliased to `docker compose`
anyways (I actually think if it wasn't, certain things break anyways).
This is just using the proper format for certain.

I went through running several docker commands and noticed no
difference, as this change shouldn't meaningfully change anything, that
is to be expected.