[Task] Update analytics db to use local.env and Remove Dynaconf from Analytics (#136)
## Summary
Fixes #107 
Fixes #115 

### Time to review: __5 mins__

## Changes proposed
* removed the `*.toml` files related to dynaconf
* removed references to Dynaconf (e.g. in Docstrings, gitignore)
* use Pydantic for loading

## Context for reviewers

> After getting feedback for
https://github.com/navapbc/simpler-grants-gov/issues/107, the consensus
was to reevaluate the way the database loader works for more uniformity.
Changes will need to be made primarily in db.py and cli.py
> 
> With the PR https://github.com/navapbc/simpler-grants-gov/pull/84 ,
the env settings for the db are stored in settings.toml. The config
settings should be updated to use the existing local.env file

## Additional information
> Screenshots, GIF demos, code examples or output to help show the
changes working as expected.