[Issue #184] Pin Python version to 3.12 + dependency updates (#185)
## Summary
Fixes #184

### Time to review: __5 mins__

## Changes proposed
Pin the Python version and document details on how to upgrade

Upgraded packages as well while I was tinkering with this

## Context for reviewers
This is mirroring changes from the template repo:
https://github.com/navapbc/template-application-flask/pull/235

Python releases new minor versions (3.12, 3.13, etc.) every year in
October. It usually takes a few weeks for all of our dependencies and
tooling to also be upgraded to the latest version, causing our builds to
break. There isn't much we can do except wait a few weeks and then do
the upgrade (assuming no breaking changes in features we use).

However, we had some of our dependencies pinned to the major version
(Python 3) so it has broken the past few years until it started working
again when the dependencies got fixed. This is just getting ahead of
that and making sure the upgrade to Python 3.13 doesn't cause any
problems.

## Additional information
This change is largely documentation as the Python version used in the
dockerfile + pyproject.toml already would have resolved to Python 3.12,
this just makes it so it won't auto-upgrade to 3.13 when that releases
in October.

The package updates are all very minor - we updated them not too long
ago, mostly just cleaning up a few things like the types-requests issue
that is no longer present.