[Issue #2033] More filters in search schema (navapbc/simpler-grants-gov#189)
Fixes #2033

Added a builder for the search schema for integer and boolean fields.

Updated the string builder to fix a small bug + allow you to specify a
pattern for the string.

This is just adding additional filters that we'll be able to search by
(once the rest of the implementation is built out). These are all fields
that fall into a "yeah, I think someone might want to narrow down by
that" group, and definitely doesn't encompass every filter we might want
to add. Mostly just wanted to get integer and boolean fields implemented
so the search logic could all get built out with actual use cases.

Added examples to the OpenAPI docs to verify these work, at least at the
request parsing phase.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>