[Issue #2184] Tidy up python and docker usage in analytics (#2185)
## Summary
Fixes #2184

### Time to review: __3 mins__

## Changes proposed
Fixed our Python version to 3.12 (dockerfile and pyproject.toml)

Removed usage of `docker-compose` command for `docker compose`

Removed version from docker-compose file (prints a warning - no longer
used)

## Context for reviewers
Similar changes have been made recently to the API. Was updating
dependencies in the folder earlier and realized these should also happen
for analytics and are pretty minimal.

The Python version change doesn't change what version of Python we use,
but I've found in past years when new versions of Python come out, a lot
of dependencies temporarily break until they've had a chance to build in
upstream repositories. This change just makes it so we can manually test
that upgrade and take care of it once things are less chaotic (note
Python releases occur in October).

`docker-compose` as a command has been deprecated for years and for the
most part wasn't being used in the makefile, but there was still one
left. I actually can't run it anymore on a newish version of docker.

The version field is deprecated and unused in the docker-compose file
and just causes a warning of `WARN[0000]
/Users/michaelchouinard/workspace/grants-equity/analytics/docker-compose.yml:
the attribute `version` is obsolete, it will be ignored, please remove
it to avoid potential confusion` to be printed, so removing that line
fixes it.