[Issue #2030] Adjust opportunity response to contain the agency name (#2182)
## Summary
Fixes #2030

### Time to review: __5 mins__

## Changes proposed
Add the agency name to the opportunity response

Make it so the agency name can be fetched in the opportunity model

## Context for reviewers
Recent work added a table for agency data, and while this won't be
populated for a bit (needs some testing in the lower envs), the tables
themselves are setup. This will attempt to fetch the agency, and if
found cause the name to be returned. This is technically how the agency
name works in the current system although the schema of the agency is
quite a bit different.

## Additional information
Works both for the data loaded into the search index as well as the data
returned by the GET /opportunity endpoint:
<img width="450" alt="Screenshot 2024-09-19 at 12 48 50 PM"
src="https://github.com/user-attachments/assets/25b82f32-b464-4e31-bcb6-6cb9d3cbb455">
<img width="494" alt="Screenshot 2024-09-19 at 12 49 02 PM"
src="https://github.com/user-attachments/assets/36d5674e-c6e3-4068-a3f2-cbdc8e544a36">

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>