[Issue #2338] Adjust set-current-opportunities script to remove status from drafts (#2349)
## Summary
Fixes #2338

### Time to review: __5 mins__

## Changes proposed
Modified the set-current-opportunities script to remove the
opportunity_status + current opportunity from any opportunity where
is_draft=True

Minor adjustment to the transformation logic to allow different
capitalization for str booleans as the Oracle test DB had a few
variations different from prod for some fields

## Context for reviewers
The set-current-opportunities script is the last step of the
transformation process which handles calculating both the current
opportunity summary (in short - figure out which forecast/non-forecast
record is "active") as well as that summaries status.

We don't want to calculate these values if the opportunity is a draft as
there is no current summary technically as it isn't "ready" yet. We
already filter drafts out of our endpoints, this just keeps the data in
the DB simpler and matching the behavior of the APIs as well.