[Issue #2405] Account for updates when last_upd_date is null (#2406)
## Summary
Fixes #2405

### Time to review: __5 mins__

## Changes proposed
Fallback to using `created_date` when `last_upd_date` is null

## Context for reviewers
The legacy Oracle database has a last_upd_date timestamp that we use for
finding updates to our system.

This field is always null when a record is first inserted, and only on
receiving at least one update will it be populated.

This means that if a record is inserted, we copy over last_upd_date=null
and then an update happens, we try to do effectively: where null < {some
timestamp} which means we find nothing when trying to update. What we
instead should do is fallback to using the created_date timestamp if the
update date is null.

## Additional information
Found this when I realized there were cases where we just weren't
processing updates. Turns out, the ELT process doesn't handle updates
pretty much at all right now (this is the common case)