Update terraform to 1.9, address deploy consistency (#2404)
### Time to review: __2 mins__

## Motivation & Context

This PR does 2 things (GASP!)

1. Prod deploys are failing because of a terraform versioning issue. As
a possible fix, I'm upgrading everything to 1.9. I'm hoping that fixes
the deploys, and if it doesn't, I shall look into this much more
closely.

2. I updated the deploy run names so its easier to tell what is running
when. This was purely to make my life easier.

## Additional information

[Here are the terraform 1.8 => 1.9
docs](https://developer.hashicorp.com/terraform/language/upgrade-guides)

The things that apply to us are...

- nothing, we don't use `import blocks`
- we also don't use `.tftest.hcl`

## Testing

Tested on this commit:
https://github.com/HHS/simpler-grants-gov/pull/2404/commits/9a70469c65301d277e4e991b6389be9385e8e318